+++ client/Cargo.toml
@@ -19,13 +19,20 @@ tokio-stream = { workspace = true }
 async-trait = { workspace = true }
 async-recursion = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
-tokio = { workspace = true, features = ["signal", "rt-multi-thread", "parking_lot"] }
+tokio = { workspace = true, features = [
+	"signal",
+	"rt-multi-thread",
+	"parking_lot",
+] }
 k256 = { workspace = true, features = ["ecdsa"] }
 sha3 = { workspace = true }
 
 # Bifrost Relayer
 br-primitives = { path = "../primitives", default-features = false }
 br-metrics = { path = "../metrics", default-features = false }
+
+# Substrate
+sc-service = { workspace = true }
 
 [dev-dependencies]
 serde_json = { workspace = true }
