+++ client/src/eth/tx/legacy_manager.rs
@@ -1,13 +1,9 @@
-use crate::eth::{
-	generate_delay, EthClient, EventMessage, LegacyGasMiddleware, TransactionManager, TxRequest,
-	DEFAULT_TX_RETRIES,
-};
+use crate::eth::{EthClient, EventMessage, LegacyGasMiddleware, TransactionManager, TxRequest};
 use async_trait::async_trait;
 use br_primitives::{
 	constants::{DEFAULT_ESCALATE_PERCENTAGE, DEFAULT_MIN_GAS_PRICE},
 	eth::ETHEREUM_BLOCK_TIME,
-	sub_display_format, INSUFFICIENT_FUNDS, NETWORK_DOES_NOT_SUPPORT_EIP1559,
-	PROVIDER_INTERNAL_ERROR,
+	sub_display_format, NETWORK_DOES_NOT_SUPPORT_EIP1559, PROVIDER_INTERNAL_ERROR,
 };
 use ethers::{
 	middleware::{
@@ -18,19 +14,21 @@ use ethers::{
 	signers::LocalWallet,
 	types::{BlockId, BlockNumber, Transaction, TransactionRequest, U256},
 };
-use std::{sync::Arc, time::Duration};
-use tokio::{
-	sync::{
-		mpsc,
-		mpsc::{UnboundedReceiver, UnboundedSender},
-	},
-	time::sleep,
+use sc_service::SpawnTaskHandle;
+use std::sync::Arc;
+use tokio::sync::{
+	mpsc,
+	mpsc::{UnboundedReceiver, UnboundedSender},
 };
+
+use super::{AsyncTransactionTask, LegacyTransactionTask};
 
 const SUB_LOG_TARGET: &str = "legacy-tx-manager";
 
-type LegacyMiddleware<T> = NonceManagerMiddleware<
-	SignerMiddleware<Arc<GasEscalatorMiddleware<Arc<Provider<T>>>>, LocalWallet>,
+pub type LegacyMiddleware<T> = Arc<
+	NonceManagerMiddleware<
+		SignerMiddleware<Arc<GasEscalatorMiddleware<Arc<Provider<T>>>>, LocalWallet>,
+	>,
 >;
 
 /// The essential task that sends legacy transactions asynchronously.
@@ -54,6 +52,8 @@ pub struct LegacyTransactionManager<T> {
 	/// If first relay transaction is stuck in mempool after waiting for this amount of time(ms),
 	/// ignore duplicate prevent logic. (default: 12s)
 	duplicate_confirm_delay: Option<u64>,
+	/// A handle for spawning transaction tasks in the service.
+	tx_spawn_handle: SpawnTaskHandle,
 }
 
 impl<T: 'static + JsonRpcClient> LegacyTransactionManager<T> {
@@ -65,6 +65,7 @@ impl<T: 'static + JsonRpcClient> LegacyTransactionManager<T> {
 		min_gas_price: Option<u64>,
 		is_initially_escalated: bool,
 		duplicate_confirm_delay: Option<u64>,
+		tx_spawn_handle: SpawnTaskHandle,
 	) -> (Self, UnboundedSender<EventMessage>) {
 		let (sender, receiver) = mpsc::unbounded_channel::<EventMessage>();
 
@@ -90,13 +91,14 @@ impl<T: 'static + JsonRpcClient> LegacyTransactionManager<T> {
 		(
 			Self {
 				client,
-				middleware,
+				middleware: Arc::new(middleware),
 				receiver,
 				is_txpool_enabled: false,
 				is_initially_escalated,
 				gas_price_coefficient,
 				min_gas_price: U256::from(min_gas_price.unwrap_or(DEFAULT_MIN_GAS_PRICE)),
 				duplicate_confirm_delay,
+				tx_spawn_handle,
 			},
 			sender,
 		)
@@ -109,16 +111,27 @@ impl<T: 'static + JsonRpcClient> TransactionManager<T> for LegacyTransactionMana
 		self.is_txpool_enabled
 	}
 
-	fn debug_mode(&self) -> bool {
-		self.client.debug_mode
-	}
-
 	fn get_client(&self) -> Arc<EthClient<T>> {
 		self.client.clone()
 	}
 
-	fn duplicate_confirm_delay(&self) -> Duration {
-		Duration::from_millis(self.duplicate_confirm_delay.unwrap_or(ETHEREUM_BLOCK_TIME * 1000))
+	fn get_spawn_handle(&self) -> SpawnTaskHandle {
+		self.tx_spawn_handle.clone()
+	}
+
+	async fn spawn_send_transaction(&self, msg: EventMessage) {
+		let spawn_handle = self.get_spawn_handle();
+		let task = LegacyTransactionTask::new(
+			self.get_client(),
+			self.middleware.clone(),
+			self.is_txpool_enabled(),
+			self.is_initially_escalated,
+			self.gas_price_coefficient,
+			self.min_gas_price,
+			self.duplicate_confirm_delay,
+		);
+		spawn_handle
+			.spawn("send_transaction", None, async move { task.try_send_transaction(msg).await });
 	}
 
 	async fn initialize(&mut self) {
@@ -188,76 +201,7 @@ impl<T: 'static + JsonRpcClient> TransactionManager<T> for LegacyTransactionMana
 				msg.metadata,
 			);
 
-			self.try_send_transaction(msg).await;
-		}
-	}
-
-	async fn try_send_transaction(&self, mut msg: EventMessage) {
-		if msg.retries_remaining == 0 {
-			return;
-		}
-
-		// sets a random delay on external chain transactions on first try
-		if msg.give_random_delay && msg.retries_remaining == DEFAULT_TX_RETRIES {
-			sleep(Duration::from_millis(generate_delay())).await;
-		}
-
-		// set transaction `from` field
-		msg.tx_request = msg.tx_request.from(self.client.address());
-
-		// estimate the gas amount to be used
-		let estimated_gas =
-			match self.middleware.estimate_gas(&msg.tx_request.to_typed(), None).await {
-				Ok(estimated_gas) => {
-					br_metrics::increase_rpc_calls(&self.client.get_chain_name());
-					U256::from(
-						(estimated_gas.as_u64() as f64 * msg.gas_coefficient.into_f64()).ceil()
-							as u64,
-					)
-				},
-				Err(error) => {
-					return self.handle_failed_gas_estimation(SUB_LOG_TARGET, msg, &error).await
-				},
-			};
-		msg.tx_request = msg.tx_request.gas(estimated_gas);
-
-		// check the txpool for transaction duplication prevention
-		if !(self.is_duplicate_relay(&msg.tx_request, msg.check_mempool).await) {
-			let mut tx = msg.tx_request.to_legacy();
-			let gas_price = self.client.get_gas_price().await;
-			if !self.is_sufficient_funds(gas_price, estimated_gas).await {
-				panic!(
-					"[{}]-[{}]-[{}] {}",
-					&self.client.get_chain_name(),
-					SUB_LOG_TARGET,
-					self.client.address(),
-					INSUFFICIENT_FUNDS,
-				);
-			}
-
-			if self.is_initially_escalated {
-				tx = tx.gas_price(
-					self.client
-						.get_gas_price_for_escalation(
-							gas_price,
-							self.gas_price_coefficient,
-							self.min_gas_price,
-						)
-						.await,
-				);
-			}
-			let result = self.middleware.send_transaction(tx, None).await;
-			br_metrics::increase_rpc_calls(&self.client.get_chain_name());
-
-			match result {
-				Ok(pending_tx) => match pending_tx.await {
-					Ok(receipt) => {
-						self.handle_success_tx_receipt(SUB_LOG_TARGET, receipt, msg.metadata)
-					},
-					Err(error) => self.handle_failed_tx_request(SUB_LOG_TARGET, msg, &error).await,
-				},
-				Err(error) => self.handle_failed_tx_request(SUB_LOG_TARGET, msg, &error).await,
-			}
+			self.spawn_send_transaction(msg).await;
 		}
 	}
 }
